refactor(permissions): migrate permission management to a new module with simplified role and permission structure
